+++ test/components/nav-drawer.spec.js
@@ -1,0 +1,24 @@
+import { shallow } from '@vue/test-utils';
+import NavDrawer from '../../src/js/components/nav-drawer.vue';
+
+// mock method.
+jest.mock('element-util', () => {
+  return {
+    addOuterListener: jest.fn(),
+  };
+});
+
+describe('#NavDrawer', () => {
+  it('mount.', () => {
+    const wrapper = shallow(NavDrawer, {
+      propsData: {
+        id: 'drawer01',
+      },
+      slots: {
+        default: `<div class="nav-drawer-body">...</div>`,
+      }
+    });
+    expect(wrapper).toBeTruthy();
+    expect(wrapper.vm.$refs.collapse).toBeTruthy();
+  });
+});